tests: Make sure that public `FuzzyHash` is `Send`
`FuzzyHash` struct (in the version 0.1.8) is marked as `!Send` due to
complicated indirection to hide implementation details but this is only a
documentation issue (actual instantiations of `FuzzyHash` are indeed `Send`).

The next release will implement a measure to mark all instantiations of
`FuzzyHash` implement all automatic traits.

Before that happens, this commit adds a new test case to make sure that
a `FuzzyHash` instantiation (`hashes::Short` in this case) is `Send`
despite being `FuzzyHash` struct is documented as `!Send`.

It is confirmed that implementing negative `!Send` to either `FuzzyHash`
struct or its internal representation results in a compilation error.